requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# ── Computer vision ───────────────────────────────────────────────────────────
opencv-python>=4.9.0
pillow>=10.2.0

# ── Object detection (YOLOv8) ────────────────────────────────────────────────
ultralytics>=8.1.0

# ── Numerical / data ─────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
scikit-learn>=1.4.0

# ── Augmentation ─────────────────────────────────────────────────────────────
albumentations>=1.3.1

# ── ONNX export / inference ───────────────────────────────────────────────────
onnx>=1.15.0
onnxruntime>=1.17.0
# GPU runtime (install separately if needed):
# onnxruntime-gpu>=1.17.0

# ── Training utilities ────────────────────────────────────────────────────────
tqdm>=4.66.0
tensorboard>=2.15.0
matplotlib>=3.8.0

# ── Config / serialisation ────────────────────────────────────────────────────
pyyaml>=6.0.1

# ── Logging ───────────────────────────────────────────────────────────────────
loguru>=0.7.2

# ── CLI / UI ──────────────────────────────────────────────────────────────────
rich>=13.7.0
